Update predict.py
One of the significant changes includes the replacement of the .pkl file/"pickle", with a new model class that has been called "MockModel", which can be seen a few lines below. This is because we were receiving an error message in regard to the file. Also, we created a new prediction rule in which readmission is likely if inpatient visits >= 2 and A1C is abnormal. This is a relatively simple rule that can be used within the predict function. Also, it allows for a more predictable outcome based on what is input by the user. As you will see in that section of the code, index 6 represents the amount of inpatient visits, while 9 is the A1C result.