Update predict.py
The Diet variable is now a part of the predict_readmission() function. The highest index is now 10 for this model (as Diet has been added). The code is also now saying that if the patient has 2 or more inpatient visits, or is on a regular diet, the likelihood for readmission increases. Lastly, a new input has been created, in that there are five diet types, with "Regular" (which is assigned the value "0") being the default if nothing is input by the user, as it can only assume the individual is not on a specialized diet.