Update predict.py
There are now three conditions that determine whether or not readmission is required. The conditions include the number of inpatient visits being greater than or equal to two, A1C being abnormal, and patients being on a regular. Also, there was a change that improved the readability by assigning the names that represent each value and their indices.  For example, inpatient_visits = data[0][6].