Update predict.py
Here, A1C is going to be prioritized over the other factors. The block of code (Which will be noted by me within the code) is going to check whether or not A1C is normal, and if it is, then all other variables are ignored, as the model is going to make the assumption that a normal A1C equates to a low risk. As a result, a no readmission statement will be returned.